pyproject.toml
[tool.poetry]
name        = 'vxGOS'
version     = '0.0.9'
description = 'Vhex operating system oriented game creation'
repository  = 'https://github.com/YannMagnin/vxGOS.git'
readme      = 'README.md'
authors     = [
    'Yann MAGNIN <yann.magnin@protonmail.com>',
]
packages    = [
    { include = 'vxsdk', from = 'vxGOS' },
]


[tool.poetry.scripts]
vxsdk   = 'vxsdk.cli:vxsdk_cli_entry'
vxboard = 'vxsdk.cli.board:vxsdk_cli_board_entry'
vxconv  = 'vxsdk.cli.converter:vxsdk_cli_converter_entry'


[tool.poetry.dependencies]
python          = '>=3.11'
click           = "8.1.*"
toml            = "0.10.*"
pillow          = "10.2.*"
types-toml      = "*"
types-pillow    = "*"

[tool.poetry.dev-dependencies]
pynvim  = "*"
pylint  = "*"
mypy    = "*"
